+++ src/cosmos/cosmosInfos.ts
@@ -1,9 +1,11 @@
 import { coreum } from './info/coreumInfo'
+import { cosmoshub } from './info/cosmoshubInfo'
 import { osmosis } from './info/osmosisInfo'
 import { thorchainrune } from './info/thorchainruneInfo'
 
 export const cosmosPlugins = {
   coreum,
+  cosmoshub,
   osmosis,
   thorchainrune
 }