+++ scripts/cosmos-ts-protos/generate-chain-json.ts
@@ -6,9 +6,17 @@ import * as fs from 'fs'
 import * as path from 'path'
 
 const chains: { [pluginId: string]: { chainName: string; url: string } } = {
+  axelar: {
+    chainName: 'axelar',
+    url: 'https://raw.githubusercontent.com/cosmos/chain-registry/master/axelar/chain.json'
+  },
   coreum: {
     chainName: 'coreum',
     url: 'https://raw.githubusercontent.com/cosmos/chain-registry/master/coreum/chain.json'
+  },
+  cosmoshub: {
+    chainName: 'cosmoshub',
+    url: 'https://raw.githubusercontent.com/cosmos/chain-registry/master/cosmoshub/chain.json'
   },
   osmosis: {
     chainName: 'osmosis',