+++ src/cosmos/cosmosInfos.ts
@@ -1,9 +1,13 @@
+import { axelar } from './info/axelarInfo'
 import { coreum } from './info/coreumInfo'
+import { cosmoshub } from './info/cosmoshubInfo'
 import { osmosis } from './info/osmosisInfo'
 import { thorchainrune } from './info/thorchainruneInfo'
 
 export const cosmosPlugins = {
+  axelar,
   coreum,
+  cosmoshub,
   osmosis,
   thorchainrune
 }